requirements.txt
Flask >=3.0.0
pandas >= 1.3.3
requests >= 2.31.0
beautifulsoup4 >= 4.12.2
google-cloud-bigquery >= 3.14.1